--- src/UniversalDeviceServiceLib/services/DeviceService.hpp
@@ -1,35 +1,0 @@
-#pragma once
-
-#include <string>
-
-#include "BaseService.hpp"
-
-class DeviceService final : public BaseService {
-protected:
-    DeviceService(IQueryExecutor* queryExecutor);
-
-public:
-    virtual ~DeviceService() = default;
-
-protected:
-    virtual void Initialize(CrowApp& app) override;
-
-    crow::response GetSettings(const std::string& idString) const;
-
-    crow::response SetSettings(const crow::request& request, const std::string& idString);
-
-    crow::response GetCommands(const std::string& idString) const;
-
-    crow::response SetCommands(const crow::request& request, const std::string& idString);
-
-    crow::response DeleteDevice(const std::string& idString);
-
-    void OnWebSocketMessage(crow::websocket::connection& connection, const std::string& data, bool is_binary);
-
-    void OnWebSocketClose(crow::websocket::connection& connection, const std::string& reason, uint16_t code);
-
-    void TimerFunction();
-
-private:
-    friend class BaseServiceExtension;
-};
